Move `sidebarNotifications` to util file
