+++ web/src/components/nav/AppSidebar/AppSidebar.tsx
@@ -64,80 +64,9 @@ import {
 import { OrganizationDropdownMenu } from "@/src/components/OrganizationDropdownMenu/OrganizationDropdownMenu";
 import { ProjectDropdownMenu } from "@/src/components/ProjectDropdownMenu/ProjectDropdownMenu";
 import { assertUnreachable } from "@/src/utils/types";
+import { SIDEBAR_NOTIFICATIONS, type SidebarNotification } from "./utils";
 
 const TWO_WEEKS_MS = 14 * 24 * 60 * 60 * 1000;
-
-type SidebarNotification = {
-  id: string;
-  title: string;
-  description: React.ReactNode;
-  createdAt?: string;
-  link?: string;
-  linkTitle?: string;
-  linkContent?: React.ReactNode;
-  ttlMs?: number;
-};
-
-const sidebarNotifications: SidebarNotification[] = [
-  {
-    id: "lw5-1",
-    title: "Launch Week: Day 1",
-    description:
-      "Run experiments inside GitHub Actions to test every PR against a Langfuse dataset.",
-    link: "https://langfuse.com/changelog/2026-05-25-experiment-ci-cd-gates",
-    linkTitle: "Learn more",
-    createdAt: "2026-05-25",
-  },
-  {
-    id: "lw5-2",
-    title: "Launch Week: Day 2",
-    description:
-      "Langfuse agent skill turns Langfuse into a headless platform to evaluate, query and instrument your application.",
-    link: "https://langfuse.com/changelog/2026-05-26-langfuse-agent-skill",
-    linkTitle: "Learn more",
-    createdAt: "2026-05-26",
-  },
-  {
-    id: "lw5-3",
-    title: "Launch Week: Day 3",
-    description: "Fast full-text search on observation I/O via the UI and API",
-    link: "https://langfuse.com/changelog/2026-05-27-clickhouse-full-text-search-fast-mode",
-    linkTitle: "Learn more",
-    createdAt: "2026-05-27",
-  },
-  {
-    id: "lw5-4",
-    title: "Launch Week: Day 4",
-    description:
-      "Code evaluators let you score observations and experiments with Python/TypeScript checks.",
-    link: "https://langfuse.com/changelog/2026-05-28-code-evaluators",
-    linkTitle: "Learn more",
-    createdAt: "2026-05-28",
-  },
-  {
-    id: "lw5-5",
-    title: "Launch Week: Day 5",
-    description:
-      "Langfuse MCP now covers observations, metrics, scores, datasets, comments, and more.",
-    link: "https://langfuse.com/changelog/2026-05-29-mcp-update",
-    linkTitle: "Learn more",
-    createdAt: "2026-05-29",
-  },
-  {
-    id: "github-star",
-    title: "Star Langfuse",
-    description:
-      "See the latest releases and help grow the community on GitHub",
-    link: "https://github.com/langfuse/langfuse",
-    linkContent: (
-      // eslint-disable-next-line @next/next/no-img-element
-      <img
-        alt="Langfuse GitHub stars"
-        src="https://img.shields.io/github/stars/langfuse/langfuse?label=langfuse&style=social"
-      />
-    ),
-  },
-];
 
 type SelfHostedPlan = Extract<Plan, "oss" | `self-hosted:${string}`>;
 
@@ -254,7 +183,7 @@ export function AppSidebar({
 }: AppSidebarProps) {
   const activeNotifications =
     notificationState.status === "visible"
-      ? sidebarNotifications.filter((notification) => {
+      ? SIDEBAR_NOTIFICATIONS.filter((notification) => {
           if (notificationState.dismissedIds.includes(notification.id)) {
             return false;
           }
@@ -456,13 +385,17 @@ function SidebarNotifications({
           </CardHeader>
           <CardContent className="px-3 pt-1.5 pb-2.5">
             {frontNotification.link &&
-              (frontNotification.linkContent ? (
+              (frontNotification.linkImage ? (
                 <Link
                   href={frontNotification.link}
                   target="_blank"
                   onClick={() => state.onLinkClick(frontNotification.id)}
                 >
-                  {frontNotification.linkContent}
+                  {/* eslint-disable-next-line @next/next/no-img-element */}
+                  <img
+                    alt={frontNotification.linkImage.alt}
+                    src={frontNotification.linkImage.src}
+                  />
                 </Link>
               ) : (
                 <Button
